Refactored estimator classes into estimators package
