+++ bhmm/estimators/bhmm_class.py
